Revert "Update sandbox (#6382)"
This reverts commit 32b1e157110812a024fd526931a4f30dda57f5f7.

+++ src/pages/registration/new_registration/UploadRegistration.tsx
@@ -69,9 +69,7 @@ export const UploadRegistration = ({ expanded, onChange }: StartRegistrationAcco
                     removeFile={() => {
                       const uppyFiles = uppy.getFiles();
                       const uppyId = uppyFiles.find((uppyFile) => uppyFile.response?.uploadURL === file.identifier)?.id;
-                      if (uppyId) {
-                        uppy.removeFile(uppyId);
-                      }
+                      uppyId && uppy.removeFile(uppyId);
                       setUploadedFiles(
                         uploadedFiles.filter((uploadedFile) => uploadedFile.identifier !== file.identifier)
                       );
